+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Better handle syntax errors in Copy/Paste Detection step in sonar-magik-plugin.
 - Fix sslr-magik-toolkit pointing to wrong Main class.
+- Add datamodel_type_dumper.magik.
 
 0.8.0 (2023-09-27)