+++ CHANGES.md
@@ -9,6 +9,7 @@
 - Fix LocalTypeReasoner error on assignment parameter.
 - Fix source name in language server diagnostics.
 - Fix language server not showing typed Magik diagnostics.
+- Fix misnamed class TypeDocTypeExistsTypeCheck --> TypeDocTypeExistsTypedCheck.
 
 0.8.0 (2023-09-27)