+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-0.8.1 (unreleased)
+0.8.1 (2023-10-15)
 
 - Better handle syntax errors in Copy/Paste Detection step in sonar-magik-plugin.
 - Fix sslr-magik-toolkit pointing to wrong Main class.